Changing return types to Objects (from primitives) in the TaskSummary interfaces (cherry picked from commit de1c4a355a2436a3fe8f7de30180ee81ad3264bb)

+++ kogito-jbpm/jbpm-human-task/jbpm-human-task-core/src/test/java/org/jbpm/services/task/TaskQueryServiceBaseTest.java
@@ -964,7 +964,7 @@ public abstract class TaskQueryServiceBaseTest extends HumanTaskServicesBaseTest
         results = ((InternalTaskService) taskService).execute(queryCmd);
         assertEquals("List of tasks", 2, results.size());
         testOrderByTaskIdAscending(results);
-        assertEquals( "Did not order when returning tasks (first task id: " + results.get(0).getId(), firstTaskid, results.get(0).getId());
+        assertEquals( "Did not order when returning tasks (first task id: " + results.get(0).getId(), firstTaskid, results.get(0).getId().longValue());
         
         // single param tests
         results = taskService.getTasksByVariousFields(workItemIds, null, null, null, null, null, null, false);
